scripts/requirements.txt
pandas
requests
pydantic
google-cloud-bigquery
google-auth
python-dotenv
notion-client
tqdm
